sort by choice in asgn2
