+++ NucleiAsgn1/src/asgn/services/CalculateTaxes.java
@@ -1,5 +1,5 @@
 package asgn.services;
 
 public interface CalculateTaxes {
-	public double calculateTax(double price);
+    double calculateTax(double price);
 }
