--- NucleiAsgn2/bin/.gitignore
@@ -1,1 +1,0 @@
-/asgn/
