+++ NucleiAsgn2/src/asgn/service/ServiceSortRecordByChoice.java
@@ -1,0 +1,10 @@
+package asgn.service;
+
+import asgn.model.Student;
+
+import java.util.Comparator;
+import java.util.List;
+
+public interface ServiceSortRecordByChoice {
+    Comparator sortStudentRecord(List<Student> studentList);
+}
